requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
Werkzeug==3.0.1
openai==1.6.1
python-dotenv==1.0.0
email-validator==2.1.0.post1
alpaca-py==0.10.0
matplotlib==3.8.2
